refactor: Change maxBodySize type from size_t to long in ClientMaxBodySizeDirective

+++ include/ClientMaxBodySizeDirective.hpp
@@ -6,7 +6,7 @@
 /*   By: sbouheni <sbouheni@student.42mulhouse.fr>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/08 18:16:05 by sbouheni          #+#    #+#             */
-/*   Updated: 2024/10/25 11:16:15 by sbouheni         ###   ########.fr       */
+/*   Updated: 2024/11/03 08:04:19 by sbouheni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,5 +34,5 @@ public:
 private:
     ClientMaxBodySizeDirective();
 
-    size_t maxBodySize;
+    long maxBodySize;
 };
